Deployment/Deploy/requirements.txt
aniso8601==8.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.8
flask-mongoengine==0.9.5
Flask-RESTful==0.3.8
Flask-WTF==0.14.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mongoengine==0.20.0
multi-key-dict==2.0.3
pbr==5.4.5
pymongo==3.10.1
python-jenkins==1.7.0
pytz==2020.1
requests==2.24.0
redis==3.5.3
six==1.15.0
urllib3==1.25.9
Werkzeug==1.0.1
WTForms==2.3.1
uWSGI==2.0.18